return only tracer provider from trace pipelines
