+++ opentelemetry-datadog/examples/datadog.rs
@@ -1,9 +1,10 @@
 use opentelemetry::{
     global,
-    trace::{Span, TraceContextExt, Tracer},
-    Key, KeyValue, Value,
+    trace::{Span, TraceContextExt, Tracer, TracerProvider},
+    InstrumentationScope, Key, KeyValue, Value,
 };
 use opentelemetry_datadog::{new_pipeline, ApiVersion};
+use opentelemetry_semantic_conventions as semcov;
 use std::thread;
 use std::time::Duration;
 
@@ -23,11 +24,17 @@ fn bar() {
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
-    let (tracer, provider) = new_pipeline()
+    let provider = new_pipeline()
         .with_service_name("trace-demo")
         .with_api_version(ApiVersion::Version05)
         .install_simple()?;
     global::set_tracer_provider(provider.clone());
+    let scope = InstrumentationScope::builder("opentelemetry-datadog")
+        .with_version(env!("CARGO_PKG_VERSION"))
+        .with_schema_url(semcov::SCHEMA_URL)
+        .with_attributes(None)
+        .build();
+    let tracer = provider.tracer_with_scope(scope);
 
     tracer.in_span("foo", |cx| {
         let span = cx.span();
